Added additional dispose checks when setting bitmaps. Fixed the fast blur method, increasing efficency. It has a slight translation error while blurring.
Separated the method to generate a filter from changing the alpha for previews / drawing. When doRefreshEffectAlpha is true, the alpha is stored for the effect surface. Previewing uses that alpha and drawing uses 0 for alpha. It's set to refresh when the plugin runs and any time a property or the effect changes.

For ease of access, holding control while left-clicking pans in addition to a middle click.